+++ labpack/banking/__init__.py
@@ -1,0 +1,4 @@
+''' a package of methods for interacting with banking services '''
+__author__ = 'rcj1492'
+__created__ = '2017.10'
+__license__ = 'MIT'